@@ -4,12 +4,12 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>Antony</groupId>
 	<artifactId>Antony</artifactId>
-	<version>7.4.2</version>
+	<version>7.5.0</version>
 
 	<properties>
-		<resteasy.version>3.5.1.Final</resteasy.version>
+		<resteasy.version>6.2.1.Final</resteasy.version>
 		<maven.compiler.release>8</maven.compiler.release>
-		<jackson.version>2.14.0-rc2</jackson.version>
+		<jackson.version>2.14.0-rc3</jackson.version>
 		<hibernate.version>5.6.5.Final</hibernate.version>
 		<hibernate.validator.version>7.0.3.Final</hibernate.validator.version>
 		<jda.version>5.0.0-alpha.21</jda.version>
@@ -29,9 +29,9 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.10.0</version>
+				<version>3.10.1</version>
 				<configuration>
-					<release>8</release>
+					<release>11</release>
 					<encoding>UTF-8</encoding>
 				</configuration>
 			</plugin>
@@ -133,6 +133,11 @@
 			<artifactId>jackson-datatype-jsr310</artifactId>
 			<version>${jackson.version}</version>
 		</dependency>
+		<dependency>
+		    <groupId>com.fasterxml.jackson.jaxrs</groupId>
+		    <artifactId>jackson-jaxrs-json-provider</artifactId>
+		    <version>${jackson.version}</version>
+		</dependency>
 		<!-- https://mvnrepository.com/artifact/org.xerial/sqlite-jdbc -->
 		<dependency>
 			<groupId>org.xerial</groupId>
@@ -207,7 +212,7 @@
 		<dependency>
 			<groupId>org.apache.logging.log4j</groupId>
 			<artifactId>log4j-core</artifactId>
-			<version>2.17.2</version>
+			<version>2.19.0</version>
 		</dependency>
 	</dependencies>
 </project>
